shit works

+++ poll_bot.py
@@ -11,6 +11,7 @@ from dotenv import load_dotenv
 from datetime import datetime, timezone, timedelta
 import re
 import json
+import base64
 
 load_dotenv()
 
@@ -24,7 +25,7 @@ BOT_SECRET = os.getenv('BOT_SECRET')
 TARGET_USER = 'Hjerneskade(Meme Of The Day)'
 
 # Replace with the specific guild name to operate in
-TARGET_GUILD_NAME = 'twlkawlkwa'
+#TARGET_GUILD_NAME = 'Beanhole™'
 
 intents = discord.Intents.default()
 intents.message_content = True
@@ -41,12 +42,11 @@ def should_log_reaction(message, user):
     now = datetime.now(timezone.utc)
     message_age = now - message.created_at
     return (message.author.name == TARGET_USER and 
-            message.guild.name == TARGET_GUILD_NAME and
-            message_age < timedelta(hours=24)
+            message_age < timedelta(hours=72)
             and user != bot.user)
 
 def should_react_to_message(message):
-    return message.author.name == TARGET_USER and message.guild.name == TARGET_GUILD_NAME
+    return message.author.name == TARGET_USER
 
 def search_filename(filename):
     pattern = r'^memeId_(?P<memeId>[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12})_visualId_(?P<visualId>[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12})_toptextId_(?P<toptextId>[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12})_bottomtextId_(?P<bottomtextId>[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}).png$'
@@ -59,8 +59,6 @@ async def log_reaction(message, reaction, user):
     emoji = reaction.emoji
     memeId = search_filename(message.attachments[0].filename)
     headers, data = prepare_vote_data(user, memeId, str(emoji)[0])
-    print(memeId)
-    print(data)
     await post_vote(bot.get_channel(reaction.message.channel.id), headers, data, True)
 
 def react_to_message(message):
@@ -71,28 +69,26 @@ def react_to_message(message):
 async def on_ready():
     print(f'Logged in as {bot.user}')
     # Find the target guild by name
-    target_guild = discord.utils.get(bot.guilds, name=TARGET_GUILD_NAME)
-    
-    if target_guild is None:
-        print(f'Guild with name "{TARGET_GUILD_NAME}" not found.')
-        return
-    print(datetime.now())
+    #target_guild = discord.utils.get(bot.guilds, name=TARGET_GUILD_NAME)
+    #
+    #if target_guild is None:
+    #    print(f'Guild with name "{TARGET_GUILD_NAME}" not found.')
+    #    return
     # Iterate over all channels in the target guild
-    for channel in target_guild.text_channels:
-        print(f'Loading recent messages from channel: {channel.name}')
-        
-        # Fetch recent messages from the channel
-        async for message in channel.history(limit=1):  # Adjust limit as needed\
-                if should_react_to_message(message):
-                    react_to_message(message)
-
-                # Store reactions if they exist
-                for reaction in message.reactions:
-                    async for user in reaction.users():
-                        if should_log_reaction(message, user):
-                            await log_reaction(message, reaction, user)
-    print(datetime.now())
-    print("done looping")
+    #for channel in target_guild.text_channels:
+    #    print(f'Loading recent messages from channel: {channel.name}')
+    #    
+    #    # Fetch recent messages from the channel
+    #    async for message in channel.history(limit=1):  # Adjust limit as needed\
+    #            if should_react_to_message(message):
+    #                react_to_message(message)
+#
+    #            # Store reactions if they exist
+    #            for reaction in message.reactions:
+    #                async for user in reaction.users():
+    #                    if should_log_reaction(message, user):
+    #                        await log_reaction(message, reaction, user)
+    
     try:
         synced = await bot.tree.sync()
         print(f'Synced {len(synced)} commands.')
@@ -110,18 +106,13 @@ async def on_reaction_add(reaction, user):
     if should_log_reaction(message, user):
         await log_reaction(message, reaction, user)
 
-@bot.command(name='get_reactions')
-async def get_reactions(ctx, message_id: int):
-    # Fetch the reactions for a given message ID
-    if message_id in reactions_data:
-        await ctx.send(f'Reactions for message {message_id}: {reactions_data[message_id]}')
-    else:
-        await ctx.send('No reactions recorded for this message ID.')
-
-# Define a new slash command to send "hello"
-@bot.tree.command(name="hello")
-async def hello(interaction: discord.Interaction):
-    await interaction.response.send_message("hello")
+#@bot.command(name='get_reactions')
+#async def get_reactions(ctx, message_id: int):
+#    # Fetch the reactions for a given message ID
+#    if message_id in reactions_data:
+#        await ctx.send(f'Reactions for message {message_id}: {reactions_data[message_id]}')
+#    else:
+#        await ctx.send('No reactions recorded for this message ID.')
 
 def format_json(json_string):
     try:
@@ -135,18 +126,9 @@ def format_json(json_string):
     except json.JSONDecodeError as e:
         return f"Error decoding JSON: {e}"
 
-@bot.tree.command(name="initiate", description="Initiate a command with an optional argument")
-@app_commands.describe(option="The optional argument, can be 'all'")
-async def initiate(interaction: discord.Interaction, option: str = None):
-    commandname = "g4-sAB.GXC"
-    if option == 'all':
-        await interaction.response.send_message(f"`initiated {commandname} with option {option}.`")
-    else:
-        await interaction.response.send_message(f"`initiated {commandname}.`")
-
-@bot.tree.command(name="create_meme", description="Create a meme")
+@bot.tree.command(name="submit_meme", description="Submit a meme")
 @app_commands.describe(visual_file="Upload the image", top_text="Top text of the meme (optional)", bottom_text="Bottom text of the meme (optional)", topics='Comma separated list of strings like so: ["Topic1","Topic2"]')
-async def create_meme(
+async def submit_meme(
     interaction: discord.Interaction,
     visual_file: discord.Attachment,
     top_text: str = None,
@@ -174,23 +156,62 @@ async def create_meme(
             "VisualFile": (visual_file.filename, file_bytes)
         }
         
-        # Make the request to the API
-        response = requests.post(API_HOST + "Memes", data=data, files=files)
+        response = requests.post(API_HOST + "Memes?renderMeme=true", data=data, files=files)
         
-        # Send the result message
         if response.status_code == 201:
-            await interaction.followup.send("Meme created successfully!\n" + "```json\n" + format_json(response.text) + "\n```")
+            file_bytes = io.BytesIO(base64.b64decode(response.json().get("renderedMeme")))
+            file_bytes.seek(0)  
+
+            await interaction.followup.send(content="Meme submitted to the database successfully!", file=discord.File(fp=file_bytes, filename="renderedMeme.png"))
         else:
             await interaction.followup.send("Failed to create meme. Status code: " + str(response.status_code))
 
     except Exception as e:
-        # Handle exceptions and ensure a response is sent
         await interaction.followup.send(f"An error occurred: {str(e)}")
 
 
-@bot.tree.command(name='create_memetext', description='Create a meme text')
-@app_commands.describe(text='The content to create a meme with', position='The position in a meme, can be "toptext" or "bottomtext"', topics='Comma separated list of strings like so: ["Topic1","Topic2"]')
-async def create_memetext(interaction: discord.Interaction, text: str, position: str, topics: str = None):
+@bot.tree.command(name="render_meme", description="Render a meme without submitting it to the database")
+@app_commands.describe(visual_file="Upload the image", top_text="Top text of the meme (optional)", bottom_text="Bottom text of the meme (optional)")
+async def render_meme(
+    interaction: discord.Interaction,
+    visual_file: discord.Attachment,
+    top_text: str = None,
+    bottom_text: str = None,
+):
+    try:
+        # Acknowledge the interaction to prevent timeout
+        await interaction.response.defer()
+
+        # Download the attachment
+        file_bytes = await visual_file.read()
+        # Prepare the data and files for the request
+        data = {
+            "TopText": top_text,
+            "BottomText": bottom_text,
+            "FileName": visual_file.filename,
+        }
+
+        files = {
+            "VisualFile": (visual_file.filename, file_bytes)
+        }
+        
+        response = requests.get(API_HOST + "Memes/Render", data=data, files=files)
+        
+        if response.status_code == 200:
+            file_bytes = io.BytesIO(response.content)
+            file_bytes.seek(0)  
+
+            await interaction.followup.send(content="Here is the rendered meme!", file=discord.File(fp=file_bytes, filename="renderedMeme.png"))
+        else:
+            await interaction.followup.send("Failed to render meme. Status code: " + str(response.status_code))
+
+    except Exception as e:
+        await interaction.followup.send(f"An error occurred: {str(e)}")
+
+
+@bot.tree.command(name='submit_memetext', description='Submit a meme text')
+@app_commands.describe(text='The content to submit a meme with', position='The position in a meme, can be "toptext" or "bottomtext"', topics='Comma separated list of strings like so: ["Topic1","Topic2"]')
+async def submit_memetext(interaction: discord.Interaction, text: str, position: str, topics: str = None):
     await interaction.response.defer()
     if(position != "toptext" and position != "bottomtext"):
         await interaction.followup.send("Choose a position of either \"bottomtext\" or \"toptext\"")
@@ -233,18 +254,21 @@ async def delete_visual(interaction: discord.Interaction, id: str):
 
 @bot.tree.command(name='delete_text', description='Delete a MemeText)')
 @app_commands.describe(id='The ID of the element to be deleted.')
-async def delete_visual(interaction: discord.Interaction, id: str):
+async def delete_text(interaction: discord.Interaction, id: str):
     await interaction.response.defer()
     await delete_element(interaction, id, "texts/")
 
 @bot.tree.command(name='delete_meme', description='Delete a Meme)')
 @app_commands.describe(id='The ID of the element to be deleted.')
-async def delete_visual(interaction: discord.Interaction, id: str):
+async def delete_meme(interaction: discord.Interaction, id: str):
     await interaction.response.defer()
     await delete_element(interaction, id, "memes/")
 
 async def delete_element(interaction, id, endpoint):
     try:
+        print("user id:", interaction.user.id)
+        if(interaction.user.id != 319532244463255552):
+            return await interaction.followup.send("You are not allowed, need more dubloons")
         response = requests.delete(API_HOST + endpoint + id)
         response.raise_for_status()  # Raise an error for bad status codes
         
@@ -298,9 +322,9 @@ def prepare_vote_data(user, elementid, votenumber):
     return headers,data
 
 
-@bot.tree.command(name="create_memevisual", description="Create a meme")
+@bot.tree.command(name="submit_memevisual", description="Submit a meme")
 @app_commands.describe(file="Upload the image", topics='Comma separated list of strings like so: ["Topic1","Topic2"]')
-async def create_memevisual(
+async def submit_memevisual(
     interaction: discord.Interaction,
     file: discord.Attachment,
     topics: str = None
